requirements.txt
Django==1.5.4
PIL==1.1.7
South==0.8.2
Unipath==1.0
bcrypt==1.0.2
cffi==0.7.2
django-admin-honeypot==0.2.5
django-braces==1.2.2
django-sslify==0.2
feedparser==5.1.3
psycopg2==2.5.1
pycparser==2.10
selenium==2.38.1 
six==1.4.1
wsgiref==0.1.2